PKG -- [fcl] Add pop, renderPop, and sendMessage wallet-util
